Fix non-free'd memory issues
Fixed two cases where memory was not freed correctly. One was in
test cases and other one was the value returned by
spindump_capture_defaultinterface(). However, the function return
type was specified to be const char*, which can't be free'd without
type casts or compiler warnings. Therefore, in order to allow memory
to be freed, changed the return type of the function to char*.